+++ src/rules/index.ts
@@ -1,4 +1,4 @@
-import { alignMockColons } from "./align_mock_colons.js";
+import { alignObjectColons } from "./align_object_colons.js";
 import { emptyFunctionRun } from "./empty_function_run.js";
 import { fenceMultilineMocks } from "./fence_multiline_mocks.js";
 import { noNullResponse } from "./no_null_response.js";
@@ -7,7 +7,7 @@ import { noVarResponse } from "./no_var_response.js";
 import type { Rule } from "./types.js";
 
 export const builtinRules: Rule[] = [
-  alignMockColons,
+  alignObjectColons,
   fenceMultilineMocks,
   emptyFunctionRun,
   noTrailingNewline,